fix: reducers
